Split SkModule and ClassInfo
